+++ AWSLex/AWSLexModel.m
@@ -29,6 +29,140 @@ NSString *const AWSLexErrorDomain = @"com.amazonaws.AWSLexErrorDomain";
 
 @end
 
+@implementation AWSLexDeleteSessionRequest
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"botAlias" : @"botAlias",
+             @"botName" : @"botName",
+             @"userId" : @"userId",
+             };
+}
+
+@end
+
+@implementation AWSLexDeleteSessionResponse
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"botAlias" : @"botAlias",
+             @"botName" : @"botName",
+             @"sessionId" : @"sessionId",
+             @"userId" : @"userId",
+             };
+}
+
+@end
+
+@implementation AWSLexDialogAction
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"fulfillmentState" : @"fulfillmentState",
+             @"intentName" : @"intentName",
+             @"message" : @"message",
+             @"messageFormat" : @"messageFormat",
+             @"slotToElicit" : @"slotToElicit",
+             @"slots" : @"slots",
+             @"types" : @"type",
+             };
+}
+
++ (NSValueTransformer *)fulfillmentStateJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"Fulfilled"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateFulfilled);
+        }
+        if ([value caseInsensitiveCompare:@"Failed"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateFailed);
+        }
+        if ([value caseInsensitiveCompare:@"ReadyForFulfillment"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateReadyForFulfillment);
+        }
+        return @(AWSLexFulfillmentStateUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexFulfillmentStateFulfilled:
+                return @"Fulfilled";
+            case AWSLexFulfillmentStateFailed:
+                return @"Failed";
+            case AWSLexFulfillmentStateReadyForFulfillment:
+                return @"ReadyForFulfillment";
+            default:
+                return nil;
+        }
+    }];
+}
+
++ (NSValueTransformer *)messageFormatJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"PlainText"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypePlainText);
+        }
+        if ([value caseInsensitiveCompare:@"CustomPayload"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeCustomPayload);
+        }
+        if ([value caseInsensitiveCompare:@"SSML"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeSsml);
+        }
+        if ([value caseInsensitiveCompare:@"Composite"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeComposite);
+        }
+        return @(AWSLexMessageFormatTypeUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexMessageFormatTypePlainText:
+                return @"PlainText";
+            case AWSLexMessageFormatTypeCustomPayload:
+                return @"CustomPayload";
+            case AWSLexMessageFormatTypeSsml:
+                return @"SSML";
+            case AWSLexMessageFormatTypeComposite:
+                return @"Composite";
+            default:
+                return nil;
+        }
+    }];
+}
+
++ (NSValueTransformer *)typesJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"ElicitIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeElicitIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ConfirmIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeConfirmIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ElicitSlot"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeElicitSlot);
+        }
+        if ([value caseInsensitiveCompare:@"Close"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeClose);
+        }
+        if ([value caseInsensitiveCompare:@"Delegate"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeDelegate);
+        }
+        return @(AWSLexDialogActionTypeUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexDialogActionTypeElicitIntent:
+                return @"ElicitIntent";
+            case AWSLexDialogActionTypeConfirmIntent:
+                return @"ConfirmIntent";
+            case AWSLexDialogActionTypeElicitSlot:
+                return @"ElicitSlot";
+            case AWSLexDialogActionTypeClose:
+                return @"Close";
+            case AWSLexDialogActionTypeDelegate:
+                return @"Delegate";
+            default:
+                return nil;
+        }
+    }];
+}
+
+@end
+
 @implementation AWSLexGenericAttachment
 
 + (NSDictionary *)JSONKeyPathsByPropertyKey {
@@ -43,6 +177,142 @@ NSString *const AWSLexErrorDomain = @"com.amazonaws.AWSLexErrorDomain";
 
 + (NSValueTransformer *)buttonsJSONTransformer {
     return [NSValueTransformer awsmtl_JSONArrayTransformerWithModelClass:[AWSLexButton class]];
+}
+
+@end
+
+@implementation AWSLexGetSessionRequest
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"botAlias" : @"botAlias",
+             @"botName" : @"botName",
+             @"userId" : @"userId",
+             };
+}
+
+@end
+
+@implementation AWSLexGetSessionResponse
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"dialogAction" : @"dialogAction",
+             @"recentIntentSummaryView" : @"recentIntentSummaryView",
+             @"sessionAttributes" : @"sessionAttributes",
+             @"sessionId" : @"sessionId",
+             };
+}
+
++ (NSValueTransformer *)dialogActionJSONTransformer {
+    return [NSValueTransformer awsmtl_JSONDictionaryTransformerWithModelClass:[AWSLexDialogAction class]];
+}
+
++ (NSValueTransformer *)recentIntentSummaryViewJSONTransformer {
+    return [NSValueTransformer awsmtl_JSONArrayTransformerWithModelClass:[AWSLexIntentSummary class]];
+}
+
+@end
+
+@implementation AWSLexIntentSummary
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"confirmationStatus" : @"confirmationStatus",
+             @"dialogActionType" : @"dialogActionType",
+             @"fulfillmentState" : @"fulfillmentState",
+             @"intentName" : @"intentName",
+             @"slotToElicit" : @"slotToElicit",
+             @"slots" : @"slots",
+             };
+}
+
++ (NSValueTransformer *)confirmationStatusJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"None"] == NSOrderedSame) {
+            return @(AWSLexConfirmationStatusNone);
+        }
+        if ([value caseInsensitiveCompare:@"Confirmed"] == NSOrderedSame) {
+            return @(AWSLexConfirmationStatusConfirmed);
+        }
+        if ([value caseInsensitiveCompare:@"Denied"] == NSOrderedSame) {
+            return @(AWSLexConfirmationStatusDenied);
+        }
+        return @(AWSLexConfirmationStatusUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexConfirmationStatusNone:
+                return @"None";
+            case AWSLexConfirmationStatusConfirmed:
+                return @"Confirmed";
+            case AWSLexConfirmationStatusDenied:
+                return @"Denied";
+            default:
+                return nil;
+        }
+    }];
+}
+
++ (NSValueTransformer *)dialogActionTypeJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"ElicitIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeElicitIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ConfirmIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeConfirmIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ElicitSlot"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeElicitSlot);
+        }
+        if ([value caseInsensitiveCompare:@"Close"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeClose);
+        }
+        if ([value caseInsensitiveCompare:@"Delegate"] == NSOrderedSame) {
+            return @(AWSLexDialogActionTypeDelegate);
+        }
+        return @(AWSLexDialogActionTypeUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexDialogActionTypeElicitIntent:
+                return @"ElicitIntent";
+            case AWSLexDialogActionTypeConfirmIntent:
+                return @"ConfirmIntent";
+            case AWSLexDialogActionTypeElicitSlot:
+                return @"ElicitSlot";
+            case AWSLexDialogActionTypeClose:
+                return @"Close";
+            case AWSLexDialogActionTypeDelegate:
+                return @"Delegate";
+            default:
+                return nil;
+        }
+    }];
+}
+
++ (NSValueTransformer *)fulfillmentStateJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"Fulfilled"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateFulfilled);
+        }
+        if ([value caseInsensitiveCompare:@"Failed"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateFailed);
+        }
+        if ([value caseInsensitiveCompare:@"ReadyForFulfillment"] == NSOrderedSame) {
+            return @(AWSLexFulfillmentStateReadyForFulfillment);
+        }
+        return @(AWSLexFulfillmentStateUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexFulfillmentStateFulfilled:
+                return @"Fulfilled";
+            case AWSLexFulfillmentStateFailed:
+                return @"Failed";
+            case AWSLexFulfillmentStateReadyForFulfillment:
+                return @"ReadyForFulfillment";
+            default:
+                return nil;
+        }
+    }];
 }
 
 @end
@@ -259,6 +529,116 @@ NSString *const AWSLexErrorDomain = @"com.amazonaws.AWSLexErrorDomain";
 
 + (NSValueTransformer *)responseCardJSONTransformer {
     return [NSValueTransformer awsmtl_JSONDictionaryTransformerWithModelClass:[AWSLexResponseCard class]];
+}
+
+@end
+
+@implementation AWSLexPutSessionRequest
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"accept" : @"accept",
+             @"botAlias" : @"botAlias",
+             @"botName" : @"botName",
+             @"dialogAction" : @"dialogAction",
+             @"sessionAttributes" : @"sessionAttributes",
+             @"userId" : @"userId",
+             };
+}
+
++ (NSValueTransformer *)dialogActionJSONTransformer {
+    return [NSValueTransformer awsmtl_JSONDictionaryTransformerWithModelClass:[AWSLexDialogAction class]];
+}
+
+@end
+
+@implementation AWSLexPutSessionResponse
+
++ (NSDictionary *)JSONKeyPathsByPropertyKey {
+	return @{
+             @"audioStream" : @"audioStream",
+             @"contentType" : @"contentType",
+             @"dialogState" : @"dialogState",
+             @"intentName" : @"intentName",
+             @"message" : @"message",
+             @"messageFormat" : @"messageFormat",
+             @"sessionAttributes" : @"sessionAttributes",
+             @"sessionId" : @"sessionId",
+             @"slotToElicit" : @"slotToElicit",
+             @"slots" : @"slots",
+             };
+}
+
++ (NSValueTransformer *)dialogStateJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"ElicitIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogStateElicitIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ConfirmIntent"] == NSOrderedSame) {
+            return @(AWSLexDialogStateConfirmIntent);
+        }
+        if ([value caseInsensitiveCompare:@"ElicitSlot"] == NSOrderedSame) {
+            return @(AWSLexDialogStateElicitSlot);
+        }
+        if ([value caseInsensitiveCompare:@"Fulfilled"] == NSOrderedSame) {
+            return @(AWSLexDialogStateFulfilled);
+        }
+        if ([value caseInsensitiveCompare:@"ReadyForFulfillment"] == NSOrderedSame) {
+            return @(AWSLexDialogStateReadyForFulfillment);
+        }
+        if ([value caseInsensitiveCompare:@"Failed"] == NSOrderedSame) {
+            return @(AWSLexDialogStateFailed);
+        }
+        return @(AWSLexDialogStateUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexDialogStateElicitIntent:
+                return @"ElicitIntent";
+            case AWSLexDialogStateConfirmIntent:
+                return @"ConfirmIntent";
+            case AWSLexDialogStateElicitSlot:
+                return @"ElicitSlot";
+            case AWSLexDialogStateFulfilled:
+                return @"Fulfilled";
+            case AWSLexDialogStateReadyForFulfillment:
+                return @"ReadyForFulfillment";
+            case AWSLexDialogStateFailed:
+                return @"Failed";
+            default:
+                return nil;
+        }
+    }];
+}
+
++ (NSValueTransformer *)messageFormatJSONTransformer {
+    return [AWSMTLValueTransformer reversibleTransformerWithForwardBlock:^NSNumber *(NSString *value) {
+        if ([value caseInsensitiveCompare:@"PlainText"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypePlainText);
+        }
+        if ([value caseInsensitiveCompare:@"CustomPayload"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeCustomPayload);
+        }
+        if ([value caseInsensitiveCompare:@"SSML"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeSsml);
+        }
+        if ([value caseInsensitiveCompare:@"Composite"] == NSOrderedSame) {
+            return @(AWSLexMessageFormatTypeComposite);
+        }
+        return @(AWSLexMessageFormatTypeUnknown);
+    } reverseBlock:^NSString *(NSNumber *value) {
+        switch ([value integerValue]) {
+            case AWSLexMessageFormatTypePlainText:
+                return @"PlainText";
+            case AWSLexMessageFormatTypeCustomPayload:
+                return @"CustomPayload";
+            case AWSLexMessageFormatTypeSsml:
+                return @"SSML";
+            case AWSLexMessageFormatTypeComposite:
+                return @"Composite";
+            default:
+                return nil;
+        }
+    }];
 }
 
 @end